+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.56] - 2024-12-03
 
 -   Adds a Vite-based React implementation. To be used as reference implementation for front-end options.
+-   Adds support for `multifactorauth` in Python backends.
 
 ## [0.0.55] - 2024-10-17